Refactoring Tools Folder (#2038)
* refactor Added new files for date and string manipulation as well as edited existing functions to fit

* refactor Added index filed which exports everything and started cleaninig up tools file

* refactor Further improved folder structure and delted duplicate functions from old tools

* refactor Moved functions from base file further

* refactor Cleared whole tools file and ordered everything in proper file

* fix Made all tests pass because im G

* fix Made all tests pass because im G

* refactor Added all tests to the appropriate places and cleared duplicate tests

* fix Removed unused vars

* fix 2 broken tests not working needed to be defined further

* fix: Moved all invidiual functions files from trading into index and made new https file

* fix: Moved fpmm/trading functions inside one file,renamed utils to math,fixed tests and made new http file

* fix: One failing test

* fix: Unused get token deleted

* refactor: Space between functions and fixed export in tools/index to include all functuons from price and liqudity
