+++ app/src/hooks/useKlerosCuration.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react'
 
 import { getLogger } from '../util/logger'
-import { waitForBlockToSync } from '../util/networks'
+import { waitForBlockToSync } from '../util/tools'
 import { KlerosCurationData, MarketMakerData, Status } from '../util/types'
 
 import { ConnectedWeb3Context } from './connectedWeb3'
