+++ app/src/services/cpk.ts
@@ -8,7 +8,6 @@ import { Transaction, verifyProxyAddress } from '../util/cpk'
 import { getLogger } from '../util/logger'
 import {
   bridgeTokensList,
-  getBySafeTx,
   getContractAddress,
   getNativeAsset,
   getTargetSafeImplementation,
@@ -16,7 +15,6 @@ import {
   getWrapToken,
   networkIds,
   pseudoNativeAssetAddress,
-  waitForBlockToSync,
 } from '../util/networks'
 import {
   calcDistributionHint,
@@ -25,6 +23,7 @@ import {
   isCToken,
   signaturesFormatted,
   waitABit,
+  waitForBlockToSync,
 } from '../util/tools'
 import { MarketData, Question, Token, TransactionStep } from '../util/types'
